Split TableOfContents, use IntersectionObserver to check which questions are in viewport
